UHF-212: Initial working service migration
