+++ config/install/language.content_settings.tpr_service.tpr_service.yml
@@ -1,0 +1,17 @@
+uuid: c75b7a66-293c-4caa-8fb8-b6e00e312b95
+langcode: en
+status: true
+dependencies:
+  module:
+    - content_translation
+    - helfi_tpr
+third_party_settings:
+  content_translation:
+    enabled: true
+    bundle_settings:
+      untranslatable_fields_hide: '0'
+id: tpr_service.tpr_service
+target_entity_type_id: tpr_service
+target_bundle: tpr_service
+default_langcode: site_default
+language_alterable: true
