+++ helfi_tpr.links.menu.yml
@@ -9,3 +9,15 @@ entity.tpr_unit.collection:
   description: List of TPR unit entities
   route_name: entity.tpr_unit.collection
   parent: system.admin_content
+
+tpr_service.settings:
+  title: TPR - Service settings
+  description: Configure TPR service entity type
+  route_name:  tpr_service.settings
+  parent: system.admin_structure
+
+entity.tpr_service.collection:
+  title: TPR - Service entities
+  description: List of TPR service entities
+  route_name: entity.tpr_service.collection
+  parent: system.admin_content
