+++ helfi_tpr.links.task.yml
@@ -24,3 +24,30 @@ tpr_unit.content_list:
   route_name: entity.tpr_unit.collection
   base_route: system.admin_content
   weight: 10
+
+tpr_service.settings:
+  title: Settings
+  route_name: tpr_service.settings
+  base_route: tpr_service.settings
+
+entity.tpr_service.view:
+  title: View
+  route_name: entity.tpr_service.canonical
+  base_route: entity.tpr_service.canonical
+
+entity.tpr_service.edit_form:
+  title: Edit
+  route_name: entity.tpr_service.edit_form
+  base_route: entity.tpr_service.canonical
+
+entity.tpr_service.delete_form:
+  title: Delete
+  route_name: entity.tpr_service.delete_form
+  base_route: entity.tpr_service.canonical
+  weight: 10
+
+tpr_service.content_list:
+  title: TPR - Service
+  route_name: entity.tpr_service.collection
+  base_route: system.admin_content
+  weight: 10
